User profile implementation
User Stories:
Profile Screen
As a user who has finished launching the application, I want to see a Profile Screen depicting my profile data.
 
As a user on the Profile Screen, I want to be able to tap on the View Purchases button to navigate to the Purchases Screen.
 
Purchases Screen
As a user on the Purchases Screen, I want to see my purchases as a list, displaying item name, purchase date, and price.
 
As a user on the Purchases Screen, I want to be able to tap on a purchase to expand the view and see additional information about that purchase.
 
As a user on the Purchases Screen, I want to be able to navigate back to the Profile Screen.